+++ lib/install/vue.rb
@@ -9,7 +9,7 @@ insert_into_file Rails.root.join("config/webpack/environment.js").to_s,
   after: "require('@rails/webpacker')\n"
 
 insert_into_file Rails.root.join("config/webpack/environment.js").to_s,
-  "environment.plugins.append('VueLoaderPlugin', new VueLoaderPlugin())\n",
+  "environment.plugins.prepend('VueLoaderPlugin', new VueLoaderPlugin())\n",
   before: "module.exports"
 
 say "Adding vue loader to config/webpack/environment.js"
@@ -18,11 +18,11 @@ insert_into_file Rails.root.join("config/webpack/environment.js").to_s,
   after: "require('vue-loader')\n"
 
 insert_into_file Rails.root.join("config/webpack/environment.js").to_s,
-  "environment.loaders.append('vue', vue)\n",
+  "environment.loaders.prepend('vue', vue)\n",
   before: "module.exports"
 
 say "Updating webpack paths to include .vue file extension"
-insert_into_file Webpacker.config.config_path, "- .vue\n".indent(4), after: /extensions:\n/
+insert_into_file Webpacker.config.config_path, "- .vue\n".indent(4), after: /\s+extensions:\n/
 
 say "Copying the example entry file to #{Webpacker.config.source_entry_path}"
 copy_file "#{__dir__}/examples/vue/hello_vue.js",