+++ package/utils/get_style_rule.js
@@ -3,10 +3,9 @@ const { resolve } = require('path')
 const devServer = require('../dev_server')
 const { nodeEnv } = require('../env')
 
-const isProduction = nodeEnv === 'production'
 const inDevServer = process.argv.find(v => v.includes('webpack-dev-server'))
 const isHMR = inDevServer && (devServer && devServer.hmr)
-const extractCSS = !isHMR || isProduction
+const extractCSS = !isHMR || nodeEnv === 'production'
 
 const styleLoader = {
   loader: 'style-loader',
@@ -45,7 +44,8 @@ const getStyleRule = (test, modules = false, preprocessors = []) => {
     use.unshift(styleLoader)
   }
 
-  return Object.assign({}, { test, use }, options)
+  // sideEffects - See https://github.com/webpack/webpack/issues/6571
+  return Object.assign({}, { test, use, sideEffects: !modules }, options)
 }
 
 module.exports = getStyleRule