+++ package/utils/helpers.js
@@ -2,6 +2,8 @@ const isObject = value => typeof value === 'object'
   && value !== null
   && (value.length === undefined || value.length === null)
 
+const isNotObject = value => !isObject(value)
+
 const isBoolean = str => /^true/.test(str) || /^false/.test(str)
 
 const isEmpty = value => value === null || value === undefined
@@ -31,15 +33,24 @@ const chdirTestApp = () => {
 
 const chdirCwd = () => process.chdir(process.cwd())
 
+const resetEnv = () => {
+  process.env = {}
+}
+
+const ensureTrailingSlash = path => (path.endsWith('/') ? path : `${path}/`)
+
 module.exports = {
   chdirTestApp,
   chdirCwd,
+  ensureTrailingSlash,
   isObject,
+  isNotObject,
   isBoolean,
   isArray,
   isEqual,
   isEmpty,
   isStrPath,
   canMerge,
-  prettyPrint
+  prettyPrint,
+  resetEnv
 }